Split Graph handling out of DetailActivity
No functionality change
